requirements.txt
torch==2.2.2
transformers==4.39.3
sentence-transformers==2.6.1
faiss-cpu==1.7.4
Pillow==10.3.0
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
streamlit==1.46.1